requirements.txt
# Conexão com PostgreSQL
psycopg2-binary==2.9.9

# Manipulação de dados
pandas==2.1.0
numpy==1.24.3

# Visualização de dados
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Jupyter
jupyter==1.0.0
notebook==6.5.4
```
